+++ lib/src/cover_crypt/cleartext_header.dart
@@ -1,8 +1,0 @@
-import 'dart:typed_data';
-
-class CleartextHeader {
-  Uint8List cleartext;
-  Uint8List metadata;
-
-  CleartextHeader(this.cleartext, this.metadata);
-}