+++ lib/src/cover_crypt/cover_crypt.dart
@@ -1,20 +1,20 @@
 import 'dart:typed_data';
 
-import 'cleartext_header.dart';
 import 'ffi.dart';
+import 'plaintext_header.dart';
 
 class CoverCryptDecryption {
   Uint8List asymmetricDecryptionKey;
 
   CoverCryptDecryption(this.asymmetricDecryptionKey);
 
-  CleartextHeader decryptWithAuthenticationData(
+  PlaintextHeader decryptWithAuthenticationData(
       Uint8List ciphertextBytes, Uint8List authenticationData) {
     return Ffi.decryptWithAuthenticationData(
         asymmetricDecryptionKey, ciphertextBytes, authenticationData);
   }
 
-  CleartextHeader decrypt(Uint8List ciphertextBytes) {
+  PlaintextHeader decrypt(Uint8List ciphertextBytes) {
     return Ffi.decrypt(asymmetricDecryptionKey, ciphertextBytes);
   }
 }