+++ test/cover_crypt_test.dart
@@ -2,11 +2,10 @@
 import 'dart:convert';
 import 'dart:io';
 
-// Package imports:
-import 'package:flutter_test/flutter_test.dart';
-
 // Project imports:
 import 'package:cloudproof/cloudproof.dart';
+// Package imports:
+import 'package:flutter_test/flutter_test.dart';
 
 import 'resources/test_vector/non_regression_test_vectors.dart';
 
@@ -16,14 +15,14 @@ final ciphertext = base64Decode(
 final key = base64Decode(
     "1gQaQuzgSmCwHBog+ADpfjOVfWC4Ue6FIEYmjvfHcQhD0MqNNL1cvyF324IIFK9w5FOwWDDKXLKifVskjCjJBwqxeGROFHknVx93TyLEnHkbH+0xqTtHF9mcb1FT0oVZCT7hecwZ4fcxqBrMVVDFFJg3nBhf9oBUrlPNd+C2wr0NVz8lwr/Ujf2wkAuVvyF65cGd1O4nt7dJolz5ouqVowQyP0WqE8mpPMnBkKX1N2g50suTLTffLbXhGXVyqhvuDQ2z/TkA/DxykKux78tGviWkwpVBVybgDw/WQ1W391oDAA6LLC4Ozgi8ow6njXSLLREOW+ezF8AAZahzKDJm1QPLNc97X1g1aASin+3Xz026j2EcamZ+X9CJZ+/05Cb7CXH4JXyapuNAEZRcu0JvnLVMkyT5aWnCe3AWalA4388HOzkc/nVozC0n+oAZJ2/RBUpMwBBHifmvn7BQNYt8Wws57rkdpmywZtW6YfzNcCdGPvKXyRZ6n+eki2KYjj46CA==");
 
-final cleartext = base64Decode("TXkgc2VjcmV0IG1lc3NhZ2U=");
+final plaintext = base64Decode("TXkgc2VjcmV0IG1lc3NhZ2U=");
 
 void main() {
   group('CoverCrypt', () {
     test('CoverCryptDecryption.decrypt', () async {
       final result = CoverCryptDecryption(key).decrypt(ciphertext);
 
-      expect(result.cleartext, equals(cleartext));
+      expect(result.plaintext, equals(plaintext));
     });
 
     test('coverCryptDecryptionWithCache.decrypt', () async {
@@ -31,7 +30,7 @@ void main() {
 
       final result = coverCryptDecryptionWithCache.decrypt(ciphertext);
 
-      expect(result.cleartext, equals(cleartext));
+      expect(result.plaintext, equals(plaintext));
     });
 
     test('nonRegressionTest', () async {