+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ---
 ## [3.0.0] - 2022-11-30
 ### Added
+- add FFI call for cover_crypt encryption
 - test vectors verification on cover_crypt
 ### Changed
 - update to cover_crypt 7.1.1 and findex 1.0